Conversational_RAG_Chatbot/requirements.txt
streamlit==1.32.0
langchain==0.1.16
langchain-community==0.0.34
langchain-core==0.1.45
langchain-groq==0.1.3
langchain-chroma==0.1.2
sentence-transformers==2.7.0
chromadb==0.4.24
pypdf==4.1.0
python-dotenv==1.0.1
uuid==1.30